Initial implementation of Bonn dataset tool
Based off TUM implementation due to documentation mentioning "The sequences are in the same format as the TUM RGB-D Dataset, so that the same evaluation tools can be used."

Regex was adapted to handle scientific values found in some of the Bonn groundtruth.txt files.

A few Bonn depth.txt files referenced depth images that do not exist, when this happens the specific file is logged to console and that depth frame is dropped.