Add support for equidistant distortion
Camera sensors (snapdragon / davis) use equidistant distortion, add support within CameraSensor including serialisation / deserialisation.

Note equidistant distortion is made up of 4 parameters but uses the existing type distortion_coefficents_t which is a typealias float[5]. This was used for consistency of the CameraSensor / Builder APIs. The final value of distortion_coefficents_t is ignored when used in an equidistant context.