Lifelong_loader and OpenLORIS dataset tools (#1)
* Comment out CUDA monitoring

* Add BundleFusion branch code

* Fix Makefile missing \ in OpenLORIS

* README: remove quick start for IROS challenge

The information of lifelong_loader and the new relocalize API is
added into proper places.

Also corrected the definition of other APIs.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* add PCL to framework/shared/CMakeLists to fix compiling issue

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* remove unused code introduced by this PR

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* fix depth scaling factors of TUM/BONN/ICL-NUIM

The [TUM file formats]
(https://vision.in.tum.de/data/datasets/rgbd-dataset/file_formats)
states that depth images are scaled by a factor of 5000. However,
the factor had been set to 1000 in slambench2 (disparity_params=0.001)
because there was a hardcoded dividing by 5 in SLAMFrame.cpp.

Now that we had removed the dividing by 5, we are fixing the scaling
factors for TUM and other TUM-format datasets.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* fix depth scaling factor of ICL

Following last commit

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* lifelong_loader: align traj only with gt of the first input

For lifelong SLAM evaluation, the estimated poses should be aligned
with groundtruth of the first data input. This commit implements this
feature by introducing AlignmentOutput::SetFreeze(), which will be
called by SLAMBenchConfigurationLifelong at the end of the first
input. Then, AlignmentOutput will stop updating the transformation
matrix.

Also added the alignment-technique (-a) param to lifelong_loader.
And fixed param explanation in benchmark_loader.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* support dataset-generator options for OpenLORIS in Makefile

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* remove SLAMBenchLibraryHelperLifelong

This was introduced to keep minimal change to SLAMBenchLibraryHelper.
But as libraries should be re-compiled anyway, it is unnecessary to
maintain a 'lifelong' lib helper other than the 'normal' one, which
would be another historical strangeness.

BTW, the getSensorUpdate() API is no longer needed by the libraries.
(since the workaround of reusing input interface in
 SLAMBenchConfigurationLifelong::AddInputInterface)

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* some minor code improvements

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* Add OpenLORIS targets to .PRECIOUS

* Set the default alignment method to "new" on all loaders

* Remove unused headers in OpenLORIS.cpp

* fix association and implement interpolation

The previous implementation of associate() made an assumption that gt traj
is 'denser' than the estimated traj. When it is violated, e.g. with the
OpenLORIS home1 data, runtime error will occur.

This commit refactored the associate() function to remove the assumption.
It also implements interpolation between two closest gt poses.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* fix groundtruth mismatching in SLAMBenchConfiguration

This is a quick workaround and I'm expecting more elegent fixings.

The implementation of GTBufferingFrameStream looks over-complicated
and confusing to me. I think it made an assumption that gt should
come first from the deserialiser, and the buffer works only once
at initilization (perhaps should call fastForward in GetNextFrame?).
And it may be possible to make a more straightforward implementation
of GetGTFrames.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* fix crash when groundtruth ends earlier than other frames

Keep calm, print nan, and carry on.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* change raw ptr to smart ptr for Metrics and refine cw management

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* refactor result saving and move it out from SLAMBenchConfigurationLifelong

So that other loaders could re-use it.

Also added saving groundtruth, to faciliate post analysis.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* OpenLORIS builder: transform gt target to d400_color

And changed sensor poses accordingly. Though nobody cares.

Require to re-build all .slam file in OpenLORIS.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* add absolute orientation error into ATEMetric

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* SLAMBenchConfiguration(Lifelong): consistent member naming

This does not change any program behavior.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* OpenLORIS.h/.cpp: format code

This does not change any program behaviors.

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* copy evaluate scripts from openloris_tools

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* Minor code improvements

Use auto when possible, change pass by value to const reference

* process gt pose properly in post analysis

Signed-off-by: Xuesong Shi <xuesong.shi@intel.com>

* Fix unused parameter error

* Remove open_vins directory

Co-authored-by: Mihai Bujanca <mihaibujanca@github.com>
Co-authored-by: Xuesong Shi <xuesong.shi@intel.com>